+++ test/src/service/tweets/tweets_service_test.dart
@@ -8,7 +8,7 @@ import 'package:twitter_api_v2/src/service/tweets/tweets_service.dart';
 import 'package:test/test.dart';
 import 'package:twitter_api_v2/twitter_api_v2.dart';
 
-import '../client_context_stubs.dart' as context;
+import '../../../mocks/client_context_stubs.dart' as context;
 
 void main() {
   test('.createTweet', () async {
